@@ -13,30 +13,33 @@
     "start": "next start"
   },
   "dependencies": {
-    "@material-ui/core": "^4.12.2",
-    "@material-ui/styles": "^4.11.4",
-    "luxon": "^1.28.0",
-    "next": "^11.0.1",
-    "next-offline": "^5.0.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.2.0",
+    "@emotion/cache": "^11.9.3",
+    "@emotion/react": "^11.9.3",
+    "@emotion/server": "^11.4.0",
+    "@emotion/styled": "^11.9.3",
+    "@mui/material": "^5.9.0",
+    "luxon": "^3.0.1",
+    "next": "^12.2.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.4.0",
     "react-scrollspy": "^3.4.3",
     "smoothscroll-polyfill": "^0.4.4"
   },
   "devDependencies": {
-    "@types/luxon": "^1.27.1",
-    "@types/node": "^14.17.6",
-    "@types/react": "^17.0.15",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-scrollspy": "^3.3.4",
+    "@babel/core": "^7.18.6",
+    "@types/luxon": "^2.3.2",
+    "@types/node": "^16.11.44",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-scrollspy": "^3.3.5",
     "@types/smoothscroll-polyfill": "^0.3.1",
-    "eslint": "^7.31.0",
-    "eslint-config-airbnb-typescript-prettier": "^4.2.0",
-    "eslint-config-next": "^11.0.1",
-    "gh-pages": "^3.2.3",
-    "prettier": "^2.3.2",
-    "typescript": "^4.3.5"
+    "eslint": "^8.19.0",
+    "eslint-config-airbnb-typescript-prettier": "^5.0.0",
+    "eslint-config-next": "^12.2.2",
+    "gh-pages": "^4.0.0",
+    "prettier": "^2.7.1",
+    "typescript": "^4.7.4"
   },
   "browserslist": [
     "> 1%",
