@@ -1,34 +1,44 @@
-import { ServerStyleSheets } from '@material-ui/styles';
-import Document, {
-  DocumentContext,
-  DocumentInitialProps,
-  Head,
-  Html,
-  Main,
-  NextScript,
-} from 'next/document';
-import React from 'react';
+import createEmotionServer from '@emotion/server/create-instance';
+import Document, { DocumentContext, Head, Html, Main, NextScript } from 'next/document';
+import createEmotionCache from '../utils/createEmotionCache';
 
 export default class extends Document {
-  static async getInitialProps(ctx: DocumentContext): Promise<DocumentInitialProps> {
-    const sheets = new ServerStyleSheets();
+  static async getInitialProps(ctx: DocumentContext) {
     const originalRenderPage = ctx.renderPage;
+
+    // You can consider sharing the same Emotion cache between all the SSR requests to speed up performance.
+    // However, be aware that it can have global side effects.
+    const cache = createEmotionCache();
+    const { extractCriticalToChunks } = createEmotionServer(cache);
 
     ctx.renderPage = () =>
       originalRenderPage({
-        enhanceApp: (App) => (props) => sheets.collect(<App {...props} />),
+        enhanceApp: (App: any) =>
+          function EnhanceApp(props) {
+            return <App emotionCache={cache} {...props} />;
+          },
       });
 
     const initialProps = await Document.getInitialProps(ctx);
+    // This is important. It prevents Emotion to render invalid HTML.
+    // See https://github.com/mui/material-ui/issues/26561#issuecomment-855286153
+    const emotionStyles = extractCriticalToChunks(initialProps.html);
+    const emotionStyleTags = emotionStyles.styles.map((style) => (
+      <style
+        data-emotion={`${style.key} ${style.ids.join(' ')}`}
+        key={style.key}
+        // eslint-disable-next-line react/no-danger
+        dangerouslySetInnerHTML={{ __html: style.css }}
+      />
+    ));
 
     return {
       ...initialProps,
-      // Styles fragment is rendered after the app and page rendering finish.
-      styles: [...React.Children.toArray(initialProps.styles), sheets.getStyleElement()],
+      emotionStyleTags,
     };
   }
 
-  render(): JSX.Element {
+  render() {
     return (
       <Html lang="en">
         <Head>
@@ -42,6 +52,8 @@ export default class extends Document {
             href="https://fonts.googleapis.com/css?family=Muli:400,800|Saira+Extra+Condensed:500,700&display=swap"
             rel="stylesheet"
           />
+          <meta name="emotion-insertion-point" content="" />
+          {(this.props as any).emotionStyleTags}
           {process.env.NODE_ENV === 'production' && (
             <>
               <script async src="https://www.googletagmanager.com/gtag/js?id=UA-41952111-1" />
